pyproject.toml
[tool.poetry]
name = "langchain-linebot"
version = "0.1.0"
description = "A LineBot implemented using LangChain, featuring conversation isolation and memory, and implemented using clean architecture for practice."
authors = ["Justin Xiao <o927416847@gmail.com>"]
license = "MIT license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cerberus = "^1.3.5"
flask = "^2.3.3"
flask-sqlalchemy = "^3.1.1"
langchain = "^0.0.301"
line-bot-sdk = "^3.5.0"
marshmallow = "^3.20.1"
pydantic = "^2.4.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.21"
werkzeug = "^2.3.7"
yfinance = "^0.2.30"
psycopg2 = "^2.9.7"
openai = "^0.28.0"
tiktoken = "^0.5.1"
google-search-results = "^2.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-cover = "^3.0.0"
pytest-coverage = "^0.0"
pytest-flask = "^1.2.0"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html
[tool.black]
line-length = 100

# https://beta.ruff.rs/docs/settings/
[tool.ruff]
line-length = 100
# https://beta.ruff.rs/docs/rules/
select = ["E", "W", "F"]
ignore = ["F401", "E501"]
# Exclude a variety of commonly ignored directories.
respect-gitignore = true
ignore-init-module-imports = true
